Remove `joi` leakage to the browser in `synthetics`
