+++ x-pack/plugins/observability_solution/synthetics/public/apps/synthetics/components/monitors_page/overview/overview/use_infinite_scroll.ts
@@ -9,7 +9,7 @@ import useThrottle from 'react-use/lib/useThrottle';
 import { useEffect, useState, MutableRefObject } from 'react';
 import useIntersection from 'react-use/lib/useIntersection';
 import { useSelector } from 'react-redux';
-import { MonitorListSortField } from '../../../../../../../common/runtime_types/monitor_management/sort_field';
+import type { MonitorListSortField } from '../../../../../../../common/runtime_types/monitor_management/sort_field';
 import { useGetUrlParams } from '../../../../hooks';
 import { selectOverviewState } from '../../../../state';
 import { MonitorOverviewItem } from '../../../../../../../common/runtime_types';
