+++ x-pack/plugins/observability_solution/synthetics/public/apps/synthetics/state/monitor_list/models.ts
@@ -7,7 +7,7 @@
 
 import { ErrorToastOptions } from '@kbn/core-notifications-browser';
 
-import { MonitorListSortField } from '../../../../../common/runtime_types/monitor_management/sort_field';
+import type { MonitorListSortField } from '../../../../../common/runtime_types/monitor_management/sort_field';
 import {
   EncryptedSyntheticsMonitor,
   FetchMonitorManagementListQueryArgs,
