+++ x-pack/plugins/observability_solution/synthetics/server/routes/monitor_cruds/monitor_validation.ts
@@ -11,7 +11,7 @@ import { formatErrors } from '@kbn/securitysolution-io-ts-utils';
 
 import { omit } from 'lodash';
 import { schema } from '@kbn/config-schema';
-import { AlertConfigSchema } from '../../../common/runtime_types/monitor_management/alert_config';
+import { AlertConfigSchema } from '../../../common/runtime_types/monitor_management/alert_config_schema';
 import { CreateMonitorPayLoad } from './add_monitor/add_monitor_api';
 import { flattenAndFormatObject } from '../../synthetics_service/project_monitor/normalizers/common_fields';
 import { PrivateLocationAttributes } from '../../runtime_types/private_locations';
